chore: move `useClicks` to `composables`
