+++ packages/client/internals/SlidesOverview.vue
@@ -5,7 +5,7 @@ import { themeVars } from '../env'
 import { breakpoints, showOverview, windowSize } from '../state'
 import { currentPage, go as goSlide, rawRoutes } from '../logic/nav'
 import { currentOverviewPage, overviewRowCount } from '../logic/overview'
-import { useFixedClicks } from '../logic/clicks'
+import { useFixedClicks } from '../composables/useClicks'
 import { getSlideClass } from '../utils'
 import SlideContainer from './SlideContainer.vue'
 import SlideWrapper from './SlideWrapper'
